JAMES-2813 extract Combinator to separate file
